+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Database functions geohistory.lawsectionnewfrom, lawsectionnewsymbol, and lawsectionnewto.
 - In database table geohistory.law, column lawpagetextbefore.
 - In database table geohistory.lawsection, constraint lawsection_check.
+- In GovernmentModel, getFolderParent function.
 
 ### Changed